common resources script
